set up club search handler
