+++ src/app/events/page.tsx
@@ -1,28 +1,23 @@
 "use client";
+import ClubSearchHandler from "@/components/landing/ClubSearchHandler";
+import SearchParamRedirect from "@/components/landing/SearchParamRedirect";
 import MapContainer from "@/components/map/MapContainer";
 import { SidebarProvider } from "@/components/ui/sidebar";
-import { useVenueStore } from "@/store/venue-store";
-import { useSearchParams } from "next/navigation";
-import React, { useEffect } from "react";
+import React, { Suspense } from "react";
 
 const EventsPage = () => {
-  const { setClubId } = useVenueStore();
-  const searchParams = useSearchParams();
-  useEffect(() => {
-    const clubId = searchParams.get("clubId");
-    if (clubId) {
-      setClubId(clubId);
-    } else {
-      setClubId(null);
-    }
-  }, [searchParams, setClubId]);
-
   return (
-    <SidebarProvider>
-      <main className=''>
-        <MapContainer />
-      </main>
-    </SidebarProvider>
+    <>
+      <Suspense fallback={null}>
+        <SearchParamRedirect />
+        <ClubSearchHandler />
+      </Suspense>
+      <SidebarProvider>
+        <main className=''>
+          <MapContainer />
+        </main>
+      </SidebarProvider>
+    </>
   );
 };
